refactor: enable real-time host-based Nibble toggling by moving UI logic to a reactive component and suppressing background message broadcast errors #86et9bk39
